protoboard: remove check for spurious wakeups
These happen occasionally because the BlackBoardOnMessageWaker is
apparently slightly racey and just triggers a loop() instead of an
interface/message-specific callback. No harm though except a bunch of
unnecessary iterations over interfaces where nothing happened.

Eventually the whole thing should however be converted to implementing
a BlackBoardInterfaceListener because that would significantly reduce
the complexity of the template code.