gologpp: relax iface field/exog param type check
We cannot require that the exog_action parameter type subsumes or is
equal to the golog++ type mapped to the interface field, because the
exog parameter might actually be a domain. So we have to allow
subsumption in both directions. The case where the exog parameter has a
restricted domain type however then brings some uncertainty because
we're blindly assigning a value to it and don't check whether it's
actually within the domain. It was like this before, but with the
overhauled type system these inconsistencies become obvious now.