navgraph: publish navgraph vis markers periodically, not only once
We may miss the visualization message, e.g., if we restart rviz late.
Even then, we want to see the markers eventually. Thus, add the navgraph
visualization thread to the WORLDSTATE hook so the markers are published
periodically. If the navgraph did not change, just publish the data from
the last iteration. Otherwise, regenerate all visualization markers.