fawkesinterface: Always auto-timestamp on write()
If auto-timestamping is enabled on an interface, write() should always
update the timestamp, even if the data haven't changed. Thus, a write()
is interpreted as the writer saying "The data in the interface (changed
or not) are valid at the time of writing". This should be more in line
with the previous semantics where simply calling any interface field
setter would trigger the "changed" flag and therefore a timestamp
update, even if the new data were identical to the old data.