execution-time-estimator-lookup: basic lookup estimator
This simple estimator connects to the mongodb instance statistics-local.
Database and collection are configurable, so are the relevant field
names that hold the skill name and the duration of a document.

Skill execution times can be provided if the database collection
contains an entry with matching skill name.

A sample is drawn randomly via mongodbs sample aggregation.