execution-time-estimator: remove const from exec time estimator interface
If a thread has to be locked within the estimator functions, then the
const qualifier becomes an issue. Since mongodb clients are not
thread-safe the lookup-estimator should lock the thread upon performing
queries to the database. This is not strictly necessary at the moment as
there should not be any concurrency issues between get_execution_time
and can_execute, but to avoid potential future problems locking should
be done anyways.