execution_time_estimator: throw exception if no provider exists
Instead of returning an optional, throw an exception if no provider
exists. The optional was used to allow some default value, but we can
also provide the default value directly from the estimators, which
allows a cleaner interface and it makes sure that all callers use the
same default value.