execution-time-estimator: switch to a priority queue
Each provider now has a priority, the list of providers is checked
starting with the provider with the highest priority. This way, we can
provide a default provider (e.g., the config provider) with a low
priority. The order of registration no longer affects the precedence,
unless two providers have the same priority.