execution-time-estimator: fix exception in config estimator w/o default
If the config estimator has no configured default, getting the execution
time estimate did not work because it always tried to read the default
from the config, even if the config provided an execution time for the
given skill. This is because the second argument of
`get_float_or_default` is not evaluated lazily, but at call site. Thus,
it is evaluated even if it is not used at all. If there was no default
configured, fetching the config value would throw an exception.

To avoid this, manually check if a config value for the given skill
exists and only read the default if that's not the case.