execution-time-estimator: match skill args with regex
Treat the skill args of the other as regex when matching. Since regex
matches are preformed over the whole range the previous behavior of
matching if the args are equal is preserved in case no special chars are
used (which should not happen in valid skill strings used for execution).