clips-executive: domain, ground-domain-effects
Till now Domain-effects were only grounded and applied after an action
EXECUTION-SUCCEED, I suppose the reason for this is to wait for
conditional effects and to save unneeded grounding work.

As the need to reason about actions effects arises, grounding the
effects every time is very computationally expensive.
Similarly to grounding preconditions, ground the effects of actions upon
their creation. Currently, the grounded effects are not used to apply
the effects when an action finals, then the grounding is still recomputed.

(Todo: make use of the grounded effects instead of recomputing them when
applying the effects)