interface: old data-change now called refresh
This basically restores the behaviour pre-PR#178, but under the name
"refresh", because that's what it truly represents. So calling any
setter on an interface field or calling mark_data_refreshed() will set
the appropriate flag, the refreshed() method will return true and the
"refresh" interface listeners will be notified.

Coming up are the changes to the rest of the code pursuant to the
rename, and then the "true" data-changed feature will be added back with
a dedicated callback.