interface: implement data_changed notification
After renaming the previous "change" callback to "refresh", the
BlackBoardInterfaceListener now regains the bb_interface_data_changed
callback, except this one will now only be called if the new data is
actually different from the old.

This also also affects some enums transmitted over the wire in the
remote blackboard code, so I'd expect that it will change the bbsync
protocol in an incompatible way.

Proceed with caution!