pddl-parser: improve expression parsing
Distinguish between different types of expressions to set a baseline for
semantic checks during parsing.

This breaks compatibility with all plugins that use the pddl-parser due
to the type change of pddl_domain::Expression.

In order to restore functionality, every occurrence of Expression
instances e has to be replaced by e.expression.

The required changes to other plugins will be made at a later stage.