buildsys: use opencv supplied library dir
The buildsys was assuming that the opencv libs can be found in the
standard library search paths. Now that opencv-3 is in the process of
being phased out of distros, that may not be true any more, so we need
to query pkgconfig for the correct library search paths as well.