cx: if a mutex is locked without a request, release it again
We may receive a mutex update that a mutex has been locked by us, but we
do not have a pending request. In this case, release the mutex again.

This should usually not occur, but if it does, it may be fatal, as we
hold a mutex that we are not aware of.