clips-executive: properly remove preconditions recursively
Previously the recursive deletion could not work as the function
argument was assumed to be an atomic precondition name.
Also, the case of a resulting empty precondition was not covered.

While fixing the issues above it prove to be more convenient to actually
change the argument from a precondition name to a fact-id/address.