interface: add a source field to interface messages
The source is the interface ID where the message was originally
enqueued, in contrast to the sender, which is always the intermediate
sender of the message. If the message is relayed, the sender changes,
while the source stays the same.

This allows us to identify where the message originally came from, e.g.,
to determine the interface ID that has exclusive control over the
skiller.