clips-executive: add run-parallel compound goal
The run-parallel goal executes all its sub-goals in parallel, starting
by selecting the ones with highest priority. Once those are DISPATCHED,
the next batch is selected.
So priorities can be used to give a head start to certain goals, e.g., to
acquire resources before other goals start or to schedule expansion
tasks.
A run-parallel stops SELECTION goals once a sub-goal is REJECTED or
FAILED, unless it is configured to continue regardless (which is done
by specifying the goal parameter 'continue-on').