execution-time-estimator-lookup: mockup regex queries
In general it would be useful to have execution time documents that can
be applied to a range of skill strings via regex patterns.

Consider for example a skill that gets a location as input and where
this location does not influence the execution duration. Without such
a regex pattern it would require a separate entry for each possible
parameter value just to cover a single execution time sample.

However, with mongodb 4.0 it is not possible to use document values as
regular expressions in a matching.
It would require mongodb 4.2 $regexMatch to build a proper query that
treats the skill arg value as regex to match against the concrete value
that should be executed.

As a workaround, only support the trivial regex ".*" which always
succeeds. In the above example scenario a single entry with location
value ".*" hence would apply to skill strings independent from their
location.