clips-pddl: add fact to store grounding information of pddl formulas
Add the pddl-grounding fact, which stores the parameter names and
values of grounded PDDL values. Each grounded formula or
predicate has a reference to the corresponding pddl-grounding
fact to model the relation explicitely. It also reduces redundant
information by decoupling parameter values from grounded predicates
and making it possible to only use the ungrounded formula facts
to rebuild the formula tree structure for grounded facts. Each
plan-action has a reference to the pddl-grounding fact that was
created based on its parameters to model the relation to the source
of the information. The grounding fact is furthermore is now used as
the reference to the precondition formula.