robot-memory: remove tests
Those tests are not run and require to actually run fawkes with a bunch
of plugins, hence they are not too useful when it comes to isolating
issues.
Since the tests are broken since the change to mongocxx driver simply
remove them.

It does not seem worth it to keep them alive.