core: (re-)introduce LockPtr and RecursiveLockPtr
LockPtr now does not take a generic Mutex as parameter, but instead is
templatized with the specific type of the mutex. This is necessary
because there is no generic Mutex class (wrapping recursive and
non-recursive mutex) anymore.

Specifically, add a class LockPtrBase with an additional template
parameter T_Mutex, which is typically either Mutex or RecursiveMutex.

Also introduce two template specializations LockPtr and RecursiveLockPtr
for a non-recursive and recursive lock pointer.