utils: extend shm_registry magic token size
The 16-character sized definition of FIREVISION_SHM_IMAGE_MAGIC_TOKEN
in shm_image.h is "FireVision Image" which will not include the null
termination.
Thus the functionality of listing SharedMemoryImageBuffers (i.e. used by
the webview plugin or fvshmem) is broken.
This extends the size so the character array and possible future
modifications fit into this size.