clips-pddl: add simple way to handle domain changes for quantified formulas
Quantified formulas need to be up to date with the objects in the domain,
thus when a new object of a matching type gets added, the corresponding
sub-tree needs to be inserted. Similarly, if an existing object gets
removed, the corresponding sub-tree needs to be removed. This
implementation provides the functionality by removing the entire grounded
formula, thus triggering a new generation of it. Later on, this
implementation should be replaced with more efficient and more targeted
one.