clips-pddl: add reference to parent to grounded-pddl-formulas
grounded-pddl-formulas had no reference to their grounded parent until now
as the tree structure of the formula was determined throught the
ungrounded formula. To check whether quantified formulas are satisfied,
it is necessary to distinguish between different grounded parent formulas
as we look into all possible groundings of the quantified parameters but
only want the correct sub-tree to be viewed as satisfied.