clips-pddl: improve runtime of quantified formula housekeeping
The housekeeping rules for quantified formulas make sure that quantified
branches where the object that the local grounding is based on is removed
are cut and new branches are added when a new object of a quantified type
is added to the wm. Improve the runtime of the housekeeping rules by
only cutting and adding specific branches and not the entire quantified
sub-tree.