pddl-robot-memory: extend template to allow arbitrary aggregations
Before, we always assumed the aggregation to be a simple `$match`. Now,
also arbitrary aggregations by extending the template language. As
before, queries that are introduced with a single `|` are treated as
queries. Additionally, introduce `||` as marker for arbitrary
aggregations. Thus, the following two template snippets are equivalent:

Short form:
```
<<#FACTS|{"relation": "domain-fact"}>> ... <</FACTS>>
```

General form:
```
<<#FACTS||{ "$match": {"relation": "domain-fact"} }>> ... <</FACTS>>
```