clips-executive: optionally allow an action to require a resource
With promises, we may have goals that have already been dispatched, but
where not all goal resources have been acquired yet, as they are still
acquired by a promising goal. If this is the case, it may make sense to
block certain actions until the required resource has actually been
acquired. In other words, we may dispatch a plan early, so the robot can
execute some prefix of the plan (e.g., moving to a location), but then
block the execution until the resource has been acquired.