clips-executive: release lock if we suddenly become the owner
We encountered a case, where the local mutex fact was saying that
another robot held the lock, hence the own mutex request was rejected.
After the rejection went through we finally got the update about the
request that was sent asynchronously (prior to the requection).
Hence the mutex fact changed from
(mutex (state LOCKED) (locked-by "other") (request NONE)
to
(mutex (state LOCKED) (locked-by "us") (request NONE)
which is of course not wanted.

Co-authored-by: Daniel Swoboda <swoboda@kbsg.rwth-aachen.de>