clips-executive: clean up mutex responses and requests on fast reject
The only problematic case that should not be cleaned up is when a mutex
has a pending response (as then we would need for a trigger). This is
already excluded through the precondition.

In particular, now the rule can clean up mutexes with request LOCK and
response NONE (which therefore did not try to lock asynchronously yet.