clips-robot-memory: check for success of clipsmm set_slot function
clipsmm and clips 6.31 now only succeed when setting a slot if the fact
was never asserted and if the given value matches the type from the slot
constraint. Otherwise the call to set_slot will not do anything.

To prevent nasty bugs, actually check for success when using this
function.