robot-memory: do not request an upsert on a mutex request
The filter doc specifies that the mutex must not be locked (which is
correct). However, that means that if the mutex is locked already, then
we will just insert a new document with the same `_id`. It seems like in
previous versions of mongodb, such a request would be silently ignored
because IDs are supposed to be unique. However, this is no longer the
case and we end up with two mutex documents with the same ID.
For this reason, we must not use upsert when requesting the mutex.

However, we still need to create the mutex document if it does not exist
at all. Therefore, before actually requesting the lock, first insert the
mutex doc if it does not exist yet.