clips-executive: combine shared object update rules to prevent dropping
If the wm-fact containing the list of objects gets updated (i.e. its
fact-id increases) without the list itself changing, neither the rule
for adding new objects, nor the rule for removing objects fires. This
means that the fact-id of the wm-fact and the id stored in the sync-map
diverge. In such a case a newly added object might be instantly deleted.
To prevent such cases, both rules are combined into one, which also
handles updating the fact-id when the contents don't change. Since
the actual update performance was computed on the RHS already, the
functionality can just be re-applied.