cmake: make buildsys configurable for usage as submodule
The idea is to only define a project if fawkes is used standalone.
Otherwise the top-level project is used along with its project source
dir as build destinations.

This is not the cleanest solution and should probably be solved
differently such that fawkes core remains a cmake project while having
adjustable destianations.