rrd: disable per default
The updated function interfaces on f40 cause incompatibility with f39.
As the package version is the same according to pkgconfig we cannot
easily distinguish the different version.
They only distinguish by a bumped release in f40 which is not reflected
in pkgconfig...
Since we do not use rrd anyways, simply disable it for now